1.0.35 (#31)
* Add --debug option

* Correctly distinguish duplicate bots in team cards

* Search for python3.10 on system

Bump rust-version to 1.65
Refactor to use let-else

* More refactors to utilize let-else

* Update story mode

* Fix some clippy warnings

More small refactors

* Send shut_down signal to match handler on GUI close

Use less for loops, perfer Rust iterators

* Update to Tauri 0.2

* Update deps

+++ src-tauri/src/stories/cmaps.rs
@@ -83,7 +83,7 @@ pub struct Bot {
     pub name: String,
     #[serde(rename = "type")]
     pub type_field: BotType,
-    pub skill: Option<f64>,
+    pub skill: Option<f32>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub path: Option<Vec<String>>,
 }
